requirements.txt
python-magic-bin; platform_system == "Windows"
python-magic; platform_system != "Windows"